+++ .env.dist
@@ -1,4 +1,5 @@
-export TERMINUS_SITE=[test-site-name]
+export TERMINUS_UPSTREAM_ID=[test-site-upstream-id]
+export TERMINUS_ORG=[test-site-org]
 export TERMINUS_TOKEN=[token]
 export TERMINUS_BIN_PATH=[path-to-local-dev-terminus-dir]/terminus
 export TERMINUS_DEPENDENCIES_BASE_DIR=[path-to-local-dev-terminus-dir]/terminus-data/dependencies-1
